requirements.txt
PyPDF2
pandas
llama-cpp-python